Split tests to make spark sessions work
