+++ src/test/scala/ai/starlake/schema/handlers/Schema2HandlerSpec.scala
@@ -38,7 +38,7 @@ import java.net.URL
 import scala.reflect.io.Directory
 import scala.util.{Failure, Success, Try}
 
-class SchemaHandlerSpec extends TestHelper {
+class Schema2HandlerSpec extends TestHelper {
 
   override def afterAll(): Unit = {
     super.afterAll()
@@ -96,325 +96,6 @@ class SchemaHandlerSpec extends TestHelper {
       .withFallback(super.testConfiguration)
   }
 
-  "Ingest Flat Locations JSON" should "produce file in accepted" in {
-    new WithSettings() {
-      // clean datasets folder
-      new Directory(new java.io.File(starlakeDatasetsPath)).deleteRecursively()
-
-      new SpecTrait(
-        sourceDomainOrJobPathname = "/sample/simple-json-locations/locations_domain.sl.yml",
-        datasetDomainName = "locations",
-        sourceDatasetPathName = "/sample/simple-json-locations/flat-locations.json"
-      ) {
-        println(s"test root is $starlakeTestRoot")
-        File(starlakeTestRoot).delete()
-        sparkSessionReset(settings)
-        cleanMetadata
-        deliverSourceDomain()
-        List(
-          "/sample/simple-json-locations/locations.sl.yml",
-          "/sample/simple-json-locations/flat_locations.sl.yml"
-        ).foreach(deliverSourceTable)
-        loadPending
-
-        // Accepted should have the same data as input
-        val acceptedDf = sparkSession
-          .sql(s"select * from $datasetDomainName.flat_locations where $getTodayCondition")
-          .drop("year", "month", "day")
-
-        val expectedAccepted =
-          sparkSession.read
-            .json(
-              getResPath("/expected/datasets/accepted/locations/locations.json")
-            )
-            .withColumn("name_upper_case", upper(col("name")))
-            .withColumn("source_file_name", lit("locations.json"))
-
-        acceptedDf.show(false)
-        expectedAccepted.show(false)
-        acceptedDf
-          .select(col("id"))
-          .except(expectedAccepted.select(col("id")))
-          .count() shouldBe 0
-        sparkSession.sql("DROP TABLE IF EXISTS locations.flat_locations").show()
-      }
-    }
-  }
-  "Ingest Locations JSON" should "produce file in accepted" in {
-    new WithSettings() {
-      new SpecTrait(
-        sourceDomainOrJobPathname = "/sample/simple-json-locations/locations_domain.sl.yml",
-        datasetDomainName = "locations",
-        sourceDatasetPathName = "/sample/simple-json-locations/locations.json"
-      ) {
-        cleanMetadata
-        deliverSourceDomain()
-        List(
-          "/sample/simple-json-locations/locations.sl.yml",
-          "/sample/simple-json-locations/flat_locations.sl.yml"
-        ).foreach(deliverSourceTable)
-        loadPending
-
-        readFileContent(
-          starlakeDatasetsPath + s"/${settings.appConfig.area.archive}/$datasetDomainName/locations.json"
-        ) shouldBe loadTextFile(
-          sourceDatasetPathName
-        )
-
-        // Accepted should have the same data as input
-        val acceptedDf =
-          sparkSession
-            .sql(s"select * from $datasetDomainName.locations where $getTodayCondition")
-            .drop("year", "month", "day")
-
-        val expectedAccepted =
-          sparkSession.read
-            .json(
-              getResPath("/expected/datasets/accepted/locations/locations.json")
-            )
-            .withColumn("name_upper_case", upper(col("name")))
-            .withColumn("source_file_name", lit("locations.json"))
-
-        acceptedDf
-          .except(expectedAccepted.select(acceptedDf.columns.map(col): _*))
-          .count() shouldBe 0
-        sparkSession.sql("DROP TABLE IF EXISTS locations.locations").show()
-      }
-    }
-  }
-
-  "Ingest Locations XML" should "produce file in accepted" in {
-    new WithSettings() {
-      new SpecTrait(
-        sourceDomainOrJobPathname = "/sample/xml/locations_domain.sl.yml",
-        datasetDomainName = "locations",
-        sourceDatasetPathName = "/sample/xml/locations.xml"
-      ) {
-        cleanMetadata
-        deliverSourceDomain()
-        deliverSourceTable("/sample/xml/locations.sl.yml")
-        loadPending
-
-        readFileContent(
-          starlakeDatasetsPath + s"/${settings.appConfig.area.archive}/$datasetDomainName/locations.xml"
-        ) shouldBe loadTextFile(
-          sourceDatasetPathName
-        )
-
-        // Accepted should have the same data as input
-        val acceptedDf = sparkSession
-          .sql(s"select * from $datasetDomainName.locations")
-          .drop("year", "month", "day")
-
-        val session = sparkSession
-
-        import session.implicits._
-
-        val (seconds, millis) =
-          acceptedDf
-            .select($"seconds", $"millis")
-            .filter($"name" like "Paris")
-            .as[(String, String)]
-            .collect()
-            .head
-
-        // We just check against the year since the test may be executed in a different time zone :)
-        seconds.substring(0, 4) shouldBe "2021"
-        millis.substring(0, 4) shouldBe "1970"
-        sparkSession.sql("DROP TABLE IF EXISTS locations.locations").show()
-      }
-    }
-  }
-  "Ingest Locations XML with XSD" should "produce file in accepted" in {
-    new WithSettings() {
-      new SpecTrait(
-        sourceDomainOrJobPathname = "/sample/xsd/locations_domain.sl.yml",
-        datasetDomainName = "locations",
-        sourceDatasetPathName = "/sample/xsd/locations.xml"
-      ) {
-        cleanMetadata
-        deliverSourceDomain()
-        deliverSourceTable("/sample/xsd/locations.sl.yml")
-
-        withSettings.deliverTestFile(
-          "/sample/xsd/locations.xsd",
-          new Path(DatasetArea.metadata, "sample/xsd/locations.xsd")
-        )
-
-        loadPending
-
-        readFileContent(
-          starlakeDatasetsPath + s"/${settings.appConfig.area.archive}/$datasetDomainName/locations.xml"
-        ) shouldBe loadTextFile(
-          sourceDatasetPathName
-        )
-
-        val acceptedDf =
-          sparkSession.sql(s"select * from $datasetDomainName.locations where $getTodayCondition")
-
-        val session = sparkSession
-
-        import session.implicits._
-
-        val (seconds, millis) =
-          acceptedDf
-            .select($"seconds", $"millis")
-            .filter($"name" like "Paris")
-            .as[(String, String)]
-            .collect()
-            .head
-
-        // We just check against the year since the test may be executed in a different time zone :)
-        seconds.substring(0, 4) shouldBe "1631"
-        millis.substring(0, 4) shouldBe "1631"
-        sparkSession.sql("DROP TABLE IF EXISTS locations.locations").show()
-      }
-    }
-  }
-
-  "Ingest schema with merge" should "succeed" in {
-    new WithSettings(
-      testConfiguration.withValue("grouped", ConfigValueFactory.fromAnyRef("false"))
-    ) {
-      new SpecTrait(
-        sourceDomainOrJobPathname = "/sample/DOMAIN.sl.yml",
-        datasetDomainName = "DOMAIN",
-        sourceDatasetPathName = "/sample/SCHEMA-VALID.dsv"
-      ) {
-        cleanMetadata
-        deliverSourceDomain()
-        List(
-          "/sample/User.sl.yml",
-          "/sample/Players.sl.yml",
-          "/sample/employee.sl.yml",
-          "/sample/complexUser.sl.yml"
-        ).foreach(deliverSourceTable)
-        getDomain("DOMAIN").foreach { domain =>
-          val result = domain.tables.map { table =>
-            table.finalName -> table.containsArrayOfRecords()
-          }
-          val expected =
-            List(("User", false), ("Players", false), ("employee", false), ("complexUser", true))
-          result should contain theSameElementsAs expected
-        }
-
-        private val validator = loadWorkflow("DOMAIN", "/sample/Players.csv")
-        validator.load(LoadConfig(accessToken = None, test = false, files = None))
-
-        deleteSourceDomains()
-        deliverSourceDomain("DOMAIN", "/sample/merge/merge-with-timestamp.sl.yml")
-        deliverSourceTable(
-          "DOMAIN",
-          "/sample/merge/PlayersTimestamp.sl.yml",
-          Some("Players.sl.yml")
-        )
-        private val validator2 = loadWorkflow("DOMAIN", "/sample/Players-merge.csv")
-        validator2.load(LoadConfig(accessToken = None, test = false, files = None))
-
-        /*
-        val accepted: Array[Row] = sparkSession.read
-          .parquet(starlakeDatasetsPath + s"/accepted/$datasetDomainName/Players")
-          .collect()
-         */
-        val accepted: Array[Row] = sparkSession
-          .sql(s"select * from $datasetDomainName.Players")
-          .collect()
-
-        // Input contains a row with an older timestamp
-        // With MergeOptions.timestamp set, that row should be ignored (the rest should be updated)
-
-        val expected: Array[Row] =
-          sparkSession.read
-            .option("header", "false")
-            .option("encoding", "UTF-8")
-            .schema(playerSchema)
-            .csv(
-              getResPath("/expected/datasets/accepted/DOMAIN/Players-merged-with-timestamp.csv")
-            )
-            .collect()
-
-        accepted should contain theSameElementsAs expected
-
-        deleteSourceDomains()
-        deliverSourceDomain("DOMAIN", "/sample/merge/simple-merge.sl.yml")
-        deliverSourceTable("DOMAIN", "/sample/merge/PlayersSimple.sl.yml", Some("Players.sl.yml"))
-
-        private val validator3 = loadWorkflow("DOMAIN", "/sample/Players-merge.csv")
-        validator3.load(LoadConfig(accessToken = None, test = false, files = None))
-
-        /*        val accepted2: Array[Row] = sparkSession.read
-          .parquet(starlakeDatasetsPath + s"/accepted/$datasetDomainName/Players")
-          .collect()
-         */
-        val accepted2: Array[Row] =
-          sparkSession.sql(s"select * from $datasetDomainName.Players").collect()
-        // Input contains a row with an older timestamp
-        // Without MergeOptions.timestamp set, the existing data should be overridden anyway
-
-        val expected2: Array[Row] =
-          sparkSession.read
-            .option("header", "false")
-            .option("encoding", "UTF-8")
-            .schema(playerSchema)
-            .csv(getResPath("/expected/datasets/accepted/DOMAIN/Players-always-override.csv"))
-            .collect()
-
-        accepted2 should contain theSameElementsAs expected2
-        deleteSourceDomain("DOMAIN", "/sample/merge/simple-merge.sl.yml")
-        sparkSession.sql("DROP TABLE IF EXISTS DOMAIN.User").show()
-        sparkSession.sql("DROP TABLE IF EXISTS DOMAIN.Players").show()
-        sparkSession.sql("DROP TABLE IF EXISTS DOMAIN.complexUser").show()
-      }
-    }
-  }
-
-  "Ingest updated schema with merge" should "produce merged results accepted" in {
-    new WithSettings() {
-      new SpecTrait(
-        sourceDomainOrJobPathname = "/sample/merge/simple-merge.sl.yml",
-        datasetDomainName = "DOMAIN",
-        sourceDatasetPathName = "/sample/Players.csv"
-      ) {
-        cleanMetadata
-        deliverSourceDomain()
-        deliverSourceTable("DOMAIN", "/sample/merge/PlayersSimple.sl.yml", Some("Players.sl.yml"))
-        loadPending
-        cleanMetadata
-
-        deliverSourceDomain("DOMAIN", "/sample/merge/merge-with-new-schema.sl.yml")
-        deliverSourceTable("DOMAIN", "/sample/merge/Players.sl.yml")
-        private val validator = loadWorkflow("DOMAIN", "/sample/merge/Players-Entitled.csv")
-        validator.load(LoadConfig(accessToken = None, test = false, files = None))
-
-        val accepted: Array[Row] = sparkSession
-          .sql(s"select PK, firstName, lastName, DOB, YEAR, MONTH from $datasetDomainName.Players")
-          .collect()
-        // Input contains a row with an older timestamp
-        // With MergeOptions.timestamp set, that row should be ignored (the rest should be updated)
-
-        accepted.foreach(println)
-        val expected =
-          sparkSession.read
-            .option("encoding", "UTF-8")
-            .schema(
-              "`PK` STRING,`firstName` STRING,`lastName` STRING,`DOB` DATE,`title` STRING,`YEAR` INT,`MONTH` INT"
-            )
-            .csv(getResPath("/expected/datasets/accepted/DOMAIN/Players-merged-entitled.csv"))
-
-        expected.createOrReplaceTempView("expected")
-        val expectedFinalDf =
-          sparkSession
-            .sql("select PK, firstName, lastName, DOB, YEAR, MONTH from expected")
-            .collect()
-        expectedFinalDf.foreach(println)
-        accepted should contain theSameElementsAs expectedFinalDf
-        sparkSession.sql("DROP TABLE IF EXISTS DOMAIN.User")
-        sparkSession.sql("DROP TABLE IF EXISTS DOMAIN.Players")
-        sparkSession.sql("DROP TABLE IF EXISTS DOMAIN.complexUser")
-      }
-    }
-  }
-
   "Ingesting data" should "adapt write based on file attributes" in {
     new WithSettings() {
       new SpecTrait(
@@ -424,7 +105,6 @@ class SchemaHandlerSpec extends TestHelper {
       ) {
         cleanMetadata
         cleanDatasets
-        TestHelper.closeSession()
         deliverSourceDomain()
         deliverSourceTable("/sample/adaptiveWrite/players.sl.yml")
         loadPending
@@ -1096,64 +776,6 @@ class SchemaHandlerSpec extends TestHelper {
         val expectedFileContent = loadTextFile("/expected/dot/acl-output.dot")
         fileContent.trim shouldBe expectedFileContent.trim
         sparkSession.sql("DROP TABLE IF EXISTS dream.segment").show()
-      }
-    }
-  }
-
-  "Ingest Dream Contact CSV with ignore" should "produce file in accepted" in {
-    new WithSettings() {
-      new SpecTrait(
-        sourceDomainOrJobPathname = "/sample/dream/dreamignore.sl.yml",
-        datasetDomainName = "dreamignore",
-        sourceDatasetPathName = "/sample/dream/OneClient_Contact_20190101_090800_008.psv"
-      ) {
-
-        cleanMetadata
-        TestHelper.closeSession()
-        cleanDatasets
-        deliverSourceDomain()
-        List("/sample/dream/client.sl.yml", "/sample/dream/segment.sl.yml").foreach(
-          deliverSourceTable
-        )
-        loadPending
-
-        readFileContent(
-          starlakeDatasetsPath + s"/archive/$datasetDomainName/OneClient_Contact_20190101_090800_008.psv"
-        ) shouldBe loadTextFile(
-          sourceDatasetPathName
-        )
-
-        val auditConnectionOptions = settings.appConfig.audit.sink.getSink().getConnection().options
-        JdbcDbUtils.withJDBCConnection(auditConnectionOptions) { conn =>
-          // drop table using jdbc statement connection conn in the lines below
-          val rs = conn
-            .createStatement()
-            .executeQuery(
-              s"select count(*) from audit.rejected where domain = '$datasetDomainName' and schema = 'client'"
-            )
-          rs.next()
-          rs.getInt(1) shouldBe 0
-        }
-
-        val acceptedDf =
-          sparkSession
-            .sql(s"select * from $datasetDomainName.client where $getTodayCondition")
-            .drop("year", "month", "day")
-            // Timezone Problem
-            .drop("customer_creation_date")
-
-        val expectedAccepted =
-          sparkSession.read
-            .schema(acceptedDf.schema)
-            .json(getResPath("/expected/datasets/accepted/dream/clientignore.json"))
-            // Timezone Problem
-            .drop("customer_creation_date")
-            .withColumn("truncated_zip_code", substring(col("zip_code"), 0, 3))
-            .withColumn("source_file_name", lit("OneClient_Contact_20190101_090800_008.psv"))
-
-        acceptedDf.columns.length shouldBe expectedAccepted.columns.length
-        acceptedDf.except(expectedAccepted).count() shouldBe 0
-        sparkSession.sql("DROP TABLE IF EXISTS dream.client").show()
       }
     }
   }
